+++ pgdb/src/lib.rs
@@ -1,5 +1,6 @@
 #![doc = include_str!("../README.md")]
 
+pub mod config;
 mod db_instance;
 mod error;